requirements.txt
aniso8601==9.0.1
cffi==1.15.0
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
flask-mongoengine==1.0.0
Flask-RESTful==0.3.9
Flask-WTF==1.0.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mongoengine==0.24.2
pycparser==2.21
PyJWT==2.4.0
pymongo==4.1.1
python-dotenv==0.20.0
pytz==2022.2.1
six==1.16.0
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
